+++ tests/test_version.py
@@ -2,4 +2,4 @@ import calandria
 
 
 def test_version():
-    assert calandria.__version__ == "2.4.1"
+    assert calandria.__version__ == "2.4.2"